requirements.txt
Flask==2.0.1
google-cloud-bigquery==3.27.0
gunicorn==20.1.0